requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
python-jose[cryptography]
SQLAlchemy
psycopg2-binary
python-dotenv